+++ tests/matlab_test_data_collectors/python_testers/cart2grid_test.py
@@ -1,6 +1,6 @@
 from unittest.mock import Mock
 
-from kwave.utils.interputils import cart2grid
+from kwave.utils.interp import cart2grid
 
 from kwave.utils.conversion import scale_time
