+++ tests/matlab_test_data_collectors/python_testers/h5io_test.py
@@ -1,4 +1,4 @@
-from kwave.utils.ioutils import *
+from kwave.utils.io import *
 import numpy as np
 from pathlib import Path
